first step towards strong typing of units; converted only time/frequency units for now, others are just typedef'ed to float; strong-typing can be turned off with flag in units_def.h; stripping away volatileness of converted variables.

+++ mcpwm_foc.cpp
@@ -475,7 +475,7 @@ namespace mcpwm_foc{
 
       // Time base configuration
       TIM_TimeBaseStructure.TIM_Period = 0xFFFFFFFF;
-      TIM_TimeBaseStructure.TIM_Prescaler = (uint16_t)((TIM12_CLOCK / TIM12_FREQ) - 1);
+      TIM_TimeBaseStructure.TIM_Prescaler = (uint16_t)(static_cast<float>(TIM12_CLOCK / TIM12_FREQ) - 1);
       TIM_TimeBaseStructure.TIM_ClockDivision = 0;
       TIM_TimeBaseStructure.TIM_CounterMode = TIM_CounterMode_Up;
       TIM_TimeBaseInit(TIM12, &TIM_TimeBaseStructure);
